14 nov

+++ netflix/src/main/java/com/avengers/netflix/repository/FilmeRepository.java
@@ -3,6 +3,6 @@ package com.avengers.netflix.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.avengers.netflix.model.Filme;
 
-public interface FilmeRepository extends JpaRepository<Filme, Integer> {
+public interface FilmeRepository extends JpaRepository<Filme, Long> {
     Filme findByTitulo (String titulo);
 }
